requirements.txt
pandas==1.5.3
selenium==4.8.2
openpyxl==3.1.1
sqlalchemy==2.0.5.post1
python-dotenv~=1.0.0
mysqlclient==2.1.1